refactor: dashboard to support mentor and student views
Split the dashboard into MentorDashboard and StudentDashboard components. The main dashboard page now conditionally renders the appropriate dashboard based on the userProfile.isMentor flag. Moved student-specific dashboard logic and UI into StudentDashboard.svelte and added a placeholder MentorDashboard.svelte.
