+++ apps/mentora/src/lib/temp.svelte.ts
@@ -1,0 +1,3 @@
+export const userProfile = $state({
+    isMentor: false,
+});
