+++ apps/mentora/src/routes/dashboard/MentorDashboard.svelte
@@ -1,0 +1,5 @@
+<svelte:head>
+    <title>Mentor Dashboard - Mentora</title>
+</svelte:head>
+
+<div>TODO</div>
